backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]==0.121.1",
    "Authlib==1.6.5",
    "PyJWT==2.9.0",
    "boto3>=1.35",
    "httpx>=0.27",

]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "pytest-timeout>=2.2",
    "httpx>=0.27",
    "ruff>=0.9",
    "mypy>=1.13",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["authlib.*", "sqlmodel.*", "alembic.*", "boto3.*", "botocore.*"]
ignore_missing_imports = true
